% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtguides-package.R
\docType{package}
\name{gtguides-package}
\alias{gtguides}
\alias{gtguides-package}
\title{gtguides: Guides for 'ggplot2' using 'gt'}
\description{
Provides additional guides for the 'ggplot2' package based on tables generated by the 'gt' package.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
