% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtlegend.R
\name{guide_gtlegend}
\alias{guide_gtlegend}
\title{Legend based on gt table}
\usage{
guide_gtlegend(
  table,
  title = waiver(),
  key_after = -Inf,
  key_colname = "",
  theme = NULL,
  position = NULL,
  override.aes = list(),
  order = 0
)
}
\arguments{
\item{table}{Either a \verb{<data.frame>} or object created by \code{\link[gt:gt]{gt()}} to
display as table.}

\item{title}{A character string indicating the title to serve as table header.
If \code{NULL}, the title is not shown. By default (\code{waiver()}), the name of the
scale object or name specified in \code{\link[ggplot2:labs]{labs()}} is used for the
title. The title set via the \code{\link[gt:tab_header]{tab_header()}} overrules
titles set via the scale or via \code{labs()}.}

\item{key_after}{Placement of the key column. When \code{-Inf} (default), the
key column becomes the first column. When \code{Inf}, it becomes the last
column. Any other values are passed on to
\code{\link[gt:cols_add]{cols_add(.after)}}.}

\item{key_colname}{A character string to use as column label for the key
column.}

\item{theme}{A \code{\link[ggplot2:theme]{theme}} object to style the guide
individually or differently from the plot's theme settings. The \code{theme}
argument in the guide overrides, and is combined with, the plot's theme.}

\item{position}{A character string indicating where the legend should be
placed relative to the plot panels.}

\item{override.aes}{A list specifying aesthetic parameters of legend keys.
See the details and examples in \code{\link[ggplot2:guide_legend]{guide_legend()}}.}

\item{order}{Positive integer less than 99 that specifies the order of this
guide among multiple guides. This controls the order in which multiple
guides are displayed, not the contents of the guide itself. If 0 (default),
the order is determined by a secret algorithm.}
}
\value{
A \verb{<Guide>} object that can be used in a non-position scale function
or in the \code{\link[ggplot2:guides]{guides()}} function.
}
\description{
Takes a table formatted by the {gt} package and adds a column of legend keys.
}
\examples{
# A standard plot
p <- ggplot(mpg, aes(displ, hwy, colour = drv)) +
  geom_point()

# Make small summary table
summary <- mpg |>
  subset(select = c(drv, cty, hwy)) |>
  aggregate(by = . ~ drv, FUN = mean)

# Using plain data.frame as guide
p + guides(colour = guide_gtlegend(summary))

# Using a formatted gt-table as guide
formatted <- summary |>
  gt::gt() |>
  gt::fmt_number(c(cty, hwy)) |>
  gt::tab_spanner("mpg", c(cty, hwy)) |>
  gt::tab_style(
    gt::cell_fill("orange"),
    gt::cells_body(cty, 2)
  ) |>
  gt::tab_style(
    gt::cell_text(color = "red"),
    gt::cells_body(drv, 3)
  ) |>
  gt::tab_header("Custom title") |>
  guide_gtlegend()
p + guides(colour = formatted)

# Using theme for style options
p + guides(colour = formatted) +
  theme(
    legend.title = element_text(face = "bold"),
    legend.text  = element_text(colour = "grey50"),
    legend.key   = element_blank(),
    legend.key.size = unit(1, "cm")
  )
}
